+++ geonode/utils.py
@@ -22,6 +22,7 @@ import base64
 import re
 import math
 import copy
+import string
 
 from urlparse import urlparse
 from collections import namedtuple
@@ -39,6 +40,12 @@ DEFAULT_TITLE=""
 DEFAULT_ABSTRACT=""
 
 INVALID_PERMISSION_MESSAGE = _("Invalid permission level.")
+
+ALPHABET = string.ascii_uppercase + string.ascii_lowercase + \
+           string.digits + '-_'
+ALPHABET_REVERSE = dict((c, i) for (i, c) in enumerate(ALPHABET))
+BASE = len(ALPHABET)
+SIGN_CHARACTER = '$'
 
 http_client = httplib2.Http()
 
@@ -462,3 +469,20 @@ def json_response(body=None, errors=None, redirect_to=None, exception=None,
        body = json.dumps(body)
    return HttpResponse(body, content_type=content_type, status=status)
 
+def num_encode(n):
+    if n < 0:
+        return SIGN_CHARACTER + num_encode(-n)
+    s = []
+    while True:
+        n, r = divmod(n, BASE)
+        s.append(ALPHABET[r])
+        if n == 0: break
+    return ''.join(reversed(s))
+
+def num_decode(s):
+    if s[0] == SIGN_CHARACTER:
+        return -num_decode(s[1:])
+    n = 0
+    for c in s:
+        n = n * BASE + ALPHABET_REVERSE[c]
+    return n
